backend/requirements.txt
openai
fastapi
uvicorn
python-dotenv
google-cloud-vision
python-multipart